feat: работа SP_CUSTOM_PAGE на уровне скетча
